feat(pesavel): fundacao backend - Quantidade decimal + produto pesavel (Fase 1)
Hortifruti/Mercado, parte 1 (modelo + backend):
- VendaItem.Quantidade: int -> decimal (numeric 10,3) pra venda por peso.
  DTOs de venda (Dto/FormDto) tambem. SNGPC/controlados seguem inteiros
  (cast). Fiscal preserva o peso (removido cast pra int).
- Produto ganha Pesavel/Unidade(UN|KG)/CodigoBalanca(PLU, indice unico
  parcial). Mapeado em ProdutoService + DTOs.
- NFC-e: qCom/qTrib com peso decimal (D4) e uCom/uTrib = unidade do
  produto (antes era 'UN' fixo + Quantidade.0000, que quebrava com peso).
- Migration AddProdutoPesavelEQuantidadeDecimal.

+++ backend/ZulexPharma.Application/DTOs/Vendas/VendaDto.cs
@@ -103,7 +103,7 @@ public class VendaItemDto
     public string ProdutoNome { get; set; } = string.Empty;
     public string? Fabricante { get; set; }
     public decimal PrecoVenda { get; set; }
-    public int Quantidade { get; set; }
+    public decimal Quantidade { get; set; }
     public decimal PercentualDesconto { get; set; }
     public decimal PercentualPromocao { get; set; }
     public decimal ValorDesconto { get; set; }
@@ -123,7 +123,7 @@ public class VendaItemFormDto
     public string ProdutoNome { get; set; } = string.Empty;
     public string? Fabricante { get; set; }
     public decimal PrecoVenda { get; set; }
-    public int Quantidade { get; set; } = 1;
+    public decimal Quantidade { get; set; } = 1;
     public decimal PercentualDesconto { get; set; }
     public decimal PercentualPromocao { get; set; }
     public decimal ValorDesconto { get; set; }
