+++ backend/ZulexPharma.Domain/Entities/VendaItem.cs
@@ -15,7 +15,8 @@ public class VendaItem
     public string ProdutoNome { get; set; } = string.Empty;
     public string? Fabricante { get; set; }
     public decimal PrecoVenda { get; set; }
-    public int Quantidade { get; set; } = 1;
+    /// <summary>Quantidade do item. Decimal pra suportar venda por peso (kg). Ver spec hortifruti.</summary>
+    public decimal Quantidade { get; set; } = 1;
     public decimal PercentualDesconto { get; set; }
     public decimal PercentualPromocao { get; set; }
     public decimal ValorDesconto { get; set; }
